+++ tests/functional/Monitor.php
@@ -206,45 +206,28 @@ class Monitor extends DbTestCase
         $monitors_id = $monitor->fields['id'];
 
         // Manually set the groups_id and groups_id_tech fields to an integer value
-        $DB->update(
-            'glpi_monitors',
+        // The update migration should mvoe all the groups to the new table directly for performance reasons (no changes to array, etc)
+        $DB->delete('glpi_groups_assets', [
+            'itemtype' => 'Monitor',
+            'items_id' => $monitors_id,
+        ]);
+        $DB->insert(
+            'glpi_groups_assets',
             [
+                'itemtype' => 'Monitor',
+                'items_id' => $monitors_id,
                 'groups_id' => 1,
-                'groups_id_tech' => 2,
+                'type' => 0 // Normal
             ],
-            [
-                'id' => $monitors_id,
-            ]
         );
-        $monitor->getFromDB($monitors_id);
-        $this->array($monitor->fields['groups_id'])->containsValues([1]);
-        $this->array($monitor->fields['groups_id_tech'])->containsValues([2]);
-
-        // Manually set the groups_id and groups_id_tech fields to 0
-        $DB->update(
-            'glpi_monitors',
+        $DB->insert(
+            'glpi_groups_assets',
             [
-                'groups_id' => 0,
-                'groups_id_tech' => 0,
+                'itemtype' => 'Monitor',
+                'items_id' => $monitors_id,
+                'groups_id' => 2,
+                'type' => 1 // Tech
             ],
-            [
-                'id' => $monitors_id,
-            ]
-        );
-        $monitor->getFromDB($monitors_id);
-        $this->array($monitor->fields['groups_id'])->isEmpty();
-        $this->array($monitor->fields['groups_id_tech'])->isEmpty();
-
-        // Manually set the groups_id and groups_id_tech fields to NULL (allowed by the DB schema)
-        $DB->update(
-            'glpi_monitors',
-            [
-                'groups_id' => null,
-                'groups_id_tech' => null,
-            ],
-            [
-                'id' => $monitors_id,
-            ]
         );
         $monitor->getFromDB($monitors_id);
         $this->array($monitor->fields['groups_id'])->isEmpty();