+++ tests/functional/NetworkEquipment.php
@@ -168,45 +168,28 @@ class NetworkEquipment extends DbTestCase
         $networkequipments_id = $networkequipment->fields['id'];
 
         // Manually set the groups_id and groups_id_tech fields to an integer value
-        $DB->update(
-            'glpi_networkequipments',
+        // The update migration should mvoe all the groups to the new table directly for performance reasons (no changes to array, etc)
+        $DB->delete('glpi_groups_assets', [
+            'itemtype' => 'NetworkEquipment',
+            'items_id' => $networkequipments_id,
+        ]);
+        $DB->insert(
+            'glpi_groups_assets',
             [
+                'itemtype' => 'NetworkEquipment',
+                'items_id' => $networkequipments_id,
                 'groups_id' => 1,
-                'groups_id_tech' => 2,
+                'type' => 0 // Normal
             ],
-            [
-                'id' => $networkequipments_id,
-            ]
         );
-        $networkequipment->getFromDB($networkequipments_id);
-        $this->array($networkequipment->fields['groups_id'])->containsValues([1]);
-        $this->array($networkequipment->fields['groups_id_tech'])->containsValues([2]);
-
-        // Manually set the groups_id and groups_id_tech fields to 0
-        $DB->update(
-            'glpi_networkequipments',
+        $DB->insert(
+            'glpi_groups_assets',
             [
-                'groups_id' => 0,
-                'groups_id_tech' => 0,
+                'itemtype' => 'NetworkEquipment',
+                'items_id' => $networkequipments_id,
+                'groups_id' => 2,
+                'type' => 1 // Tech
             ],
-            [
-                'id' => $networkequipments_id,
-            ]
-        );
-        $networkequipment->getFromDB($networkequipments_id);
-        $this->array($networkequipment->fields['groups_id'])->isEmpty();
-        $this->array($networkequipment->fields['groups_id_tech'])->isEmpty();
-
-        // Manually set the groups_id and groups_id_tech fields to NULL (allowed by the DB schema)
-        $DB->update(
-            'glpi_networkequipments',
-            [
-                'groups_id' => null,
-                'groups_id_tech' => null,
-            ],
-            [
-                'id' => $networkequipments_id,
-            ]
         );
         $networkequipment->getFromDB($networkequipments_id);
         $this->array($networkequipment->fields['groups_id'])->isEmpty();