+++ tests/functional/Peripheral.php
@@ -111,45 +111,28 @@ class Peripheral extends DbTestCase
         $peripherals_id = $peripheral->fields['id'];
 
         // Manually set the groups_id and groups_id_tech fields to an integer value
-        $DB->update(
-            'glpi_peripherals',
+        // The update migration should mvoe all the groups to the new table directly for performance reasons (no changes to array, etc)
+        $DB->delete('glpi_groups_assets', [
+            'itemtype' => 'Peripheral',
+            'items_id' => $peripherals_id,
+        ]);
+        $DB->insert(
+            'glpi_groups_assets',
             [
+                'itemtype' => 'Peripheral',
+                'items_id' => $peripherals_id,
                 'groups_id' => 1,
-                'groups_id_tech' => 2,
+                'type' => 0 // Normal
             ],
-            [
-                'id' => $peripherals_id,
-            ]
         );
-        $peripheral->getFromDB($peripherals_id);
-        $this->array($peripheral->fields['groups_id'])->containsValues([1]);
-        $this->array($peripheral->fields['groups_id_tech'])->containsValues([2]);
-
-        // Manually set the groups_id and groups_id_tech fields to 0
-        $DB->update(
-            'glpi_peripherals',
+        $DB->insert(
+            'glpi_groups_assets',
             [
-                'groups_id' => 0,
-                'groups_id_tech' => 0,
+                'itemtype' => 'Peripheral',
+                'items_id' => $peripherals_id,
+                'groups_id' => 2,
+                'type' => 1 // Tech
             ],
-            [
-                'id' => $peripherals_id,
-            ]
-        );
-        $peripheral->getFromDB($peripherals_id);
-        $this->array($peripheral->fields['groups_id'])->isEmpty();
-        $this->array($peripheral->fields['groups_id_tech'])->isEmpty();
-
-        // Manually set the groups_id and groups_id_tech fields to NULL (allowed by the DB schema)
-        $DB->update(
-            'glpi_peripherals',
-            [
-                'groups_id' => null,
-                'groups_id_tech' => null,
-            ],
-            [
-                'id' => $peripherals_id,
-            ]
         );
         $peripheral->getFromDB($peripherals_id);
         $this->array($peripheral->fields['groups_id'])->isEmpty();