+++ tests/functional/Printer.php
@@ -240,45 +240,28 @@ class Printer extends DbTestCase
         $printers_id = $printer->fields['id'];
 
         // Manually set the groups_id and groups_id_tech fields to an integer value
-        $DB->update(
-            'glpi_printers',
+        // The update migration should mvoe all the groups to the new table directly for performance reasons (no changes to array, etc)
+        $DB->delete('glpi_groups_assets', [
+            'itemtype' => 'Printer',
+            'items_id' => $printers_id,
+        ]);
+        $DB->insert(
+            'glpi_groups_assets',
             [
+                'itemtype' => 'Printer',
+                'items_id' => $printers_id,
                 'groups_id' => 1,
-                'groups_id_tech' => 2,
+                'type' => 0 // Normal
             ],
-            [
-                'id' => $printers_id,
-            ]
         );
-        $printer->getFromDB($printers_id);
-        $this->array($printer->fields['groups_id'])->containsValues([1]);
-        $this->array($printer->fields['groups_id_tech'])->containsValues([2]);
-
-        // Manually set the groups_id and groups_id_tech fields to 0
-        $DB->update(
-            'glpi_printers',
+        $DB->insert(
+            'glpi_groups_assets',
             [
-                'groups_id' => 0,
-                'groups_id_tech' => 0,
+                'itemtype' => 'Printer',
+                'items_id' => $printers_id,
+                'groups_id' => 2,
+                'type' => 1 // Tech
             ],
-            [
-                'id' => $printers_id,
-            ]
-        );
-        $printer->getFromDB($printers_id);
-        $this->array($printer->fields['groups_id'])->isEmpty();
-        $this->array($printer->fields['groups_id_tech'])->isEmpty();
-
-        // Manually set the groups_id and groups_id_tech fields to NULL (allowed by the DB schema)
-        $DB->update(
-            'glpi_printers',
-            [
-                'groups_id' => null,
-                'groups_id_tech' => null,
-            ],
-            [
-                'id' => $printers_id,
-            ]
         );
         $printer->getFromDB($printers_id);
         $this->array($printer->fields['groups_id'])->isEmpty();